--- Correcao/ModeloDeQuestao.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<prova>
-<questao tipo="ME">
-	<enunciado>
-	Quem descobriu o Brasil?
-	</enunciado>
-	<alternativa correta="true">Cabral</alternativa>
-	<alternativa correta="false">Tiradentes</alternativa>
-	<alternativa correta="false">João Sem Braço</alternativa>
-	<alternativa correta="false">Sun-Tzu</alternativa>
-</questao>
-</prova>
