--- Correcao/ProvaOnline.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
-  <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1.0"/>
-  <title>Correção Prova Online</title>
-
-  <!-- CSS  
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <link href="C:\Users\Alexandre\Desktop\portal-educacao/styles/css/materialize.css" type="text/css" rel="stylesheet" media="screen,projection"/>
-  <link href="C:\Users\Alexandre\Desktop\portal-educacao/styles/css/style.css" type="text/css" rel="stylesheet" media="screen,projection"/>
-	-->
-	<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <link href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.97.7/css/materialize.min.css" type="text/css" rel="stylesheet">
-  <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-  <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.97.7/js/materialize.min.js"></script>
-
-</head>
-
-<body>
-
-	<nav class="light-blue darken-4" role="navigation">
-     	<div class="nav-wrapper container"><a id="logo-container" href="#" class="brand-logo">Logo</a>
-      		<ul class="right hide-on-med-and-down">
-        		<li><a href="#">Navbar Link</a></li>
-      		</ul>
-
-      		<ul id="nav-mobile" class="side-nav">
-        		<li><a href="#">Navbar Link</a></li>
-      		</ul>
-      		<a href="#" data-activates="nav-mobile" class="button-collapse"><i class="material-icons">menu</i></a>
-    	</div>
-  	</nav>
-
-  	<div class="container">
-  		<br><br>
-  		
-  	</div>
-
-</body>
-
-</html>
