--- Correcao/RespostasCertas.txt
@@ -1,2 +1,0 @@
-1) Quinze
-2) Sessenta e quatro
