+++ Correcao/Web/NovoXMLProva.xml
@@ -1,0 +1,53 @@
+<prova>
+		<valor>12</valor>
+    <questao tipo="ME" aleatorio="true">
+        <dificuldade>1</dificuldade>
+        <materia>Matematica</materia>
+			<conteudo>função 2 grau</conteudo>
+        <enunciado>Uma fábrica produz casacos de determinado modelo. O preço de um desses casacos é de R$200,00, quando são vendidos 200 *casacos.
+
+    O gerente da fábrica, a partir de uma pesquisa, verificou que, para cada desconto de R$2,00 no preço de cada casaco, o número de casacos *vendidos aumenta de 5.
+
+    A maior arrecadação possível com a venda dos casacos acontecerá se a fábrica vender cada casaco por um valor, em reais, pertencente ao    intervalo</enunciado>
+         <alternativa correta="false">[105,125[</alternativa>
+         <alternativa correta="true">[105,145[</alternativa>
+         <alternativa correta="false">[105,165[</alternativa>
+         <alternativa correta="false">[105,185[</alternativa>
+     </questao>
+     <questao tipo="VF" aleatorio="true">
+         <dificuldade>2</dificuldade>
+         <materia>Geografia</materia>
+         <conteudo>Regioes</conteudo>
+         <enunciado>Marque (V) para verdadeiro ou (F) para falso </enunciado>
+         <alternativa correta="false">São Paulo e Mato Grosso estão inseridos na região sudeste.</alternativa>
+         <alternativa correta="false">Minas Gerais e Bahia fazem parte da região sudeste.</alternativa>
+         <alternativa correta="false">Amazonas e Tocantins fazem parte da região Nordeste</alternativa>
+         <alternativa correta="false">A região sudoeste é a mais nova região do Brasil desde 2012. </alternativa>
+         <alternativa correta="false">O núcleo industrial do Brasil está na região Centro-Oeste.</alternativa>
+         <alternativa correta="false"> A Amazônia faz parte da região nordeste.</alternativa>
+         <alternativa correta="true"> O setor mais industrializado do Brasil se encontra em Mato Grosso.</alternativa>
+         <alternativa correta="true">O Nordeste possui a maioria dos estados na costa do país. </alternativa>
+         <alternativa correta="false">Toda a região centro-oeste não tem contato com qualquer oceano.</alternativa>
+         <alternativa correta="false">A região sul e centro-oeste possuem três estados.</alternativa>
+     </questao>
+     
+    <questao tipo="aberta">
+    <dificuldade>3</dificuldade>
+        <materia>Literatura</materia>
+        <conteudo>Dom Casmurro</conteudo>
+    <enunciado>(FUVEST / GV) O meu fim evidente era atar as duas pontas da vida, e restaurar na velhice a adolescência. Pois, senhor, não *
+    consegui recompor o que foi nem o que fui.
+    É o que diz o narrador no segundo capítulo do romance Dom Casmurro. Afinal por que não teria ele alcançado o seu intento?
+    </enunciado>
+    </questao>
+	<questao tipo="ME" aleatorio="true">
+        <dificuldade>2</dificuldade>
+        <materia>Física</materia>
+			<conteudo>Eletromagnetismo</conteudo>
+        <enunciado>Regra da mão direita</enunciado>
+         <alternativa correta="false">Não</alternativa>
+         <alternativa correta="false">Sim</alternativa>
+         <alternativa correta="true">Nunca</alternativa>
+         <alternativa correta="false">Talvez</alternativa>
+     </questao>
+</prova>
